Initially implement extracting
Extract operation is an operation which takes the field and applies
proper shifting, according to the place in the field, e.g., having
a one-byte bit-field group, consisting 3 fields:

    X X X Y Y Y Y Z

extract() for field 'X' will return:

    X X X 0 0 0 0 0

or for field 'Y' will return:

    0 0 0 Y Y Y Y 0

This implementation works in such a way that for each of the field a
shift is calculated and put to an array, which can be accessed by the
index of the field. The fields are indexed in the same order as
forwarded to the Bitfield<> template instantiation.