Implement and test little endian serializing
I had to create extract_as_big_endian() function, because calling
extract() with fold expression previously, would be wrong, because
little endianness would be improperly done. It would be done on each
field separately, but should be done after all the fields are collected
to a single register.

So to explain it simpler; what I did was:

    serialize()
    {
        return (extract<Fields.id>() | ... | 0);
    }

This works for big endian, but doesn't for little endian, because for
each extract() call, the field had little-endian shift applied.

Performing bitwise OR on fields extracted in such a way, will make it
wrong.