Add negative compile time tests to sanitize ID
This commit adds:

* CMake helper function: CreateCompileTimeNegativeTest(), which expects
the test name, the body, being C++ code to compile, and the expected
regex to be found in the output message from the compiler.
* Two negative test cases to verify whether wrong ID error is properly
outputted to the user, when using at() or extract() with wrong Field Id.

For an example: When one creates Bitfield with Field{.id = 1} and
Field{.id = 2}, and uses an Id not specified during creation, e.g.:

    at<3>();
    extract<3>();

Then, a proper compile error will be generated: "Field ID not found".

We make use of CTest PASS_REGEX feature. When a specific regex is found
in the test output, then the test is considered passing, even when the
error code of the command is non-zero.
Within the test, we invoke compilation of an executable, which is
excluded from the default build (which is not build by default when running
'make all', 'make' or 'cmake --build .'). The executable contains the
piece of code passed to the CreateCompileTimeNegativeTest() function, as
the body, being the C++ code. A source file is generated during CMake's
Generate stage, and then this file is an input for the executable.