Static assert when the whole bitfielg group not occupied
When the bitfields accumulated bit size is not equal to the bit size of
the underlying type, then this is an error. It means that each bit
within the group, must be occupied.
A test for that assertion has been added.
One test case has been fixed, because it didn't fullfil this condition.