Add portability test with Clang13
This adds a test, enabled with JUNGLES_BITFIELD_ENABLE_PORTABILITY_CHECK
CMake cache variable, which will download Clang13 for ubuntu20.04, and
create a test, which will run the complete build, but with Clang13
compiler, instead of default local compiler.
My default compiler is GCC, so this is why Clang is a portability test.
This test has a special label "Ubuntu20", to differentiate it from the
rest basic portable tests. The label is introduced due to high times of
execution of the test.